requirements.txt
streamlit
google-generativeai
python-dotenv
sec-api
pymupdf
plotly
pandas